Update registration-schedule-job.https.html to reflect the spec
Previously, the test expectation didn't comply with the spec because it
expected there's an installing worker even if register() actually didn't
install a new worker.
This CL is to update the expectation to wait for each register() call
resolved one by one. This should work because each job must not be
coalesced so the states should be updated correctly when those promises
are resolved.

Bug: 979593
Change-Id: Ib3874e7e15ebb0dfa5608cc2ff4d3484cffd72b2